+++ src/Page2Feed.Core/Services/FeedService.cs
@@ -6,9 +6,9 @@ using HtmlAgilityPack;
 using NLog;
 using Page2Feed.Core.Model;
 using Page2Feed.Core.Model.Atom;
-using Page2Feed.Web.Services.Interfaces;
+using Page2Feed.Core.Services.Interfaces;
 
-namespace Page2Feed.Web.Services
+namespace Page2Feed.Core.Services
 {
 
     public class FeedService : IFeedService
