+++ src/Page2Feed.Web/Controllers/HomeController.cs
@@ -10,8 +10,8 @@ using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using NLog;
 using Page2Feed.Core.Model.Atom;
+using Page2Feed.Core.Services.Interfaces;
 using Page2Feed.Web.Models;
-using Page2Feed.Web.Services.Interfaces;
 
 namespace Page2Feed.Web.Controllers
 {
